[Refactor] `dropOverlay` implemented in the `DragAndDropExtension`
