+++ src/editor/contrib/blockHandleExtension/blockHandleButton.ts
@@ -1,7 +1,5 @@
 import { Button, IButtonOptions } from "src/base/browser/basic/button/button";
 
-
-
 export class BlockHandleButton extends Button {
 
     constructor(opts: IButtonOptions) {
